+++ src/pyrite/physics/collider_component.py
@@ -6,9 +6,10 @@ from weakref import WeakKeyDictionary
 
 from pymunk import ShapeFilter
 
+from ..constants import COMPONENT_TYPE
 from ..events import OnSeparate, OnTouch, WhileTouching
-from .physics_service import PhysicsService, COMPONENT_TYPE
 from .rigidbody_component import RigidbodyComponent
+from ..services import PhysicsService
 from ..types import Component
 
 if TYPE_CHECKING:
@@ -62,7 +63,7 @@ class ColliderComponent(Component):
             collision_shape.collision_type = COMPONENT_TYPE
             collision_shape.body = body
             collision_shape.filter = self.filter
-            PhysicsService.space.add(collision_shape)
+            PhysicsService.add_collider(self)
 
         self.OnTouch = OnTouch(self)
         """
