+++ src/pyrite/physics/physics_system.py
@@ -1,15 +1,12 @@
 from __future__ import annotations
 
-import math
 from typing import Any, TYPE_CHECKING
 
-import pymunk
-
 from ..types import System
-from .physics_service import PhysicsService
+
 from .collider_component import ColliderComponent
 from ..transform import TransformComponent
-from ..services import TransformService
+from ..services import PhysicsService
 
 if TYPE_CHECKING:
     from pymunk import (
@@ -62,58 +59,12 @@ class PhysicsSystem(System):
 
     def __init__(self, physics_mult=1, enabled=True, order_index=0) -> None:
         super().__init__(enabled, order_index)
-        PhysicsService.comp_handler.post_solve = post_solve
-        PhysicsService.comp_handler.separate = separate
+        PhysicsService.set_component_handlers(post_solve=post_solve, separate=separate)
         self.physics_mult = physics_mult
 
     def const_update(self, timestep: float) -> None:
 
-        self.sync_bodies_to_transforms()
-
         PhysicsService.step(timestep * self.physics_mult)
 
     def update(self, delta_time: float) -> None:
-        self.sync_transforms_to_bodies()
-
-    def sync_bodies_to_transforms(self):
-        """
-        Takes all bodies with a TransformComponent and sets their position and rotation
-        to match.
-        """
-        for body, key_object in PhysicsService.bodies.items():
-            if not (transform := TransformComponent.get(key_object)):
-                continue
-            if not transform.is_dirty():
-                continue
-            vel = body.velocity
-            ang_vel = body.angular_velocity
-            pos = transform.world_position
-            body.position = (pos.x, pos.y)
-            rot = math.radians(transform.world_rotation)
-            body.angle = rot
-            body.velocity = vel
-            body.angular_velocity = ang_vel
-            PhysicsService.space.reindex_shapes_for_body(body)
-
-    def sync_transforms_to_bodies(self):
-        """
-        Takes all TransformComponents with a rigidbody and updates their position and
-        rotation with the new calculations.
-        """
-        for body, key_object in PhysicsService.bodies.items():
-            if (
-                not (transform := TransformComponent.get(key_object))
-                or body.is_sleeping
-                or body.body_type == pymunk.Body.STATIC
-            ):
-                # No adjustments to sleeping, static, or transformless objects
-                continue
-            # TODO calculate an expected interpolation value?
-            new_pos = transform.world_position.lerp(body.position, 0.1)
-            angle_between = (
-                (math.degrees(body.angle) - transform.world_rotation) + 180
-            ) % 360 - 180
-            new_rot = angle_between / 2
-
-            TransformService.set_world_position(transform, new_pos)
-            TransformService.set_world_rotation(transform, new_rot)
+        PhysicsService.sync_transforms_to_bodies(TransformComponent)
